# Conflicts:
#	src/main/java/com/example/demo/common/repository/CategoryRepository.java
#	src/main/java/com/example/demo/faq/controller/FAQController.java
#	src/main/java/com/example/demo/faq/repository/FAQRepository.java
#	src/main/java/com/example/demo/faq/service/FAQService.java

+++ src/main/java/com/example/demo/common/repository/CategoryRepository.java
@@ -3,9 +3,7 @@ package com.example.demo.common.repository;
 import com.example.demo.common.domain.CategoryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.Optional;
-
-public interface CategoryRepository extends JpaRepository<CategoryEntity,Integer> {
+public interface CategoryRepository extends JpaRepository<CategoryEntity, Long> {
 
 
     Optional<Object> findByName(String category);
